requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
langchain==0.3.1
langgraph==0.2.28
langchain-google-genai==2.0.0
pandas==2.2.2